requirements.txt
apprise
gspread
oauth2client
pandas
python-dotenv
spotipy